requirements.txt
numpy>=2.0,<2.2
pandas>=2.2,<2.3
yfinance
requests
pyyaml
python-dateutil
pytz
nltk
google-generativeai>=0.8.0
python-dotenv
lxml 
html5lib
                 # VADER용 (최초 실행 시 vader_lexicon 다운로드 필요)
# 선택: 더 정교한 감성분석을 원할 때만
# transformers
# torch